- Feito CRUD Encomendas
- Ajustes de ID nos FormRequests

- Ajuste menu Encomendas no Front

+++ app/Http/Controllers/Admin/ResidenceController.php
@@ -143,6 +143,9 @@ class ResidenceController extends Controller
      */
     public function destroy(Residence $residence)
     {
+        if ($residence->orders() || $residence->users())
+            return redirect()->back();
+
         $residence->delete();
 
         return redirect()->route('admin.residences.index');
